requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1

# Vector Store
qdrant-client==1.7.3

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Twilio WhatsApp
twilio==8.11.1

# AI/ML
google-generativeai==0.3.2
sentence-transformers==2.3.1
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
tenacity==8.2.3